Update README.md
Hi! I have made stylistic and language changes to the README.md file from the perspective of an Editor. I have not made too many changes to the informal language. Please let me know if any of it is not clear or needs more work and I'll be glad to help!